Refactored form folders
