+++ libs/application/templates/national-agency-for-children-and-families/child-protection-notification/src/forms/adultPersonal/draftForm/index.ts
@@ -1,13 +1,13 @@
 import { buildForm } from '@island.is/application/core'
 import { FormModes } from '@island.is/application/types'
 
-import { conclusionSection } from '../shared/conclusionSection'
+import { conclusionSection } from '../../shared/conclusionSection'
 import { overviewSection } from './overviewSection'
 import { firstSection } from './firstSection'
 import { secondSection } from './secondSection'
 
-export const AdultPersonalForm = buildForm({
-  id: 'AdultPersonalForm',
+export const AdultPersonalDraftForm = buildForm({
+  id: 'AdultPersonalDraftForm',
   mode: FormModes.DRAFT,
   children: [firstSection, secondSection, overviewSection, conclusionSection],
 })
