+++ libs/application/templates/national-agency-for-children-and-families/child-protection-notification/src/forms/adultProcuration/draftForm/memmSection/educationSubSection.ts
@@ -5,8 +5,8 @@ import {
   buildTextField,
   buildDescriptionField,
 } from '@island.is/application/core'
-import { memmMessages } from '../../../lib/messages'
-import { isDayCareProvider, isSchoolType } from '../../../utils/conditionUtils'
+import { memmMessages } from '../../../../lib/messages'
+import { isDayCareProvider, isSchoolType } from '../../../../utils/conditionUtils'
 
 export const educationSubSection = buildSubSection({
   id: 'memmEducationSubSection',
