+++ libs/application/templates/national-agency-for-children-and-families/child-protection-notification/src/forms/adultProcuration/draftForm/memmSection/index.ts
@@ -1,6 +1,6 @@
 import { buildSection } from '@island.is/application/core'
-import { memmMessages } from '../../../lib/messages'
-import { showMemmSection } from '../../../utils/conditionUtils'
+import { memmMessages } from '../../../../lib/messages'
+import { showMemmSection } from '../../../../utils/conditionUtils'
 import { cultureSubSection } from './cultureSubSection'
 import { educationSubSection } from './educationSubSection'
 import { receptionSubSection } from './receptionSubSection'
