+++ libs/application/templates/national-agency-for-children-and-families/child-protection-notification/src/forms/adultProcuration/draftForm/memmSection/receptionSubSection.ts
@@ -5,9 +5,9 @@ import {
   buildRadioField,
   buildSubSection,
 } from '@island.is/application/core'
-import { memmMessages } from '../../../lib/messages'
-import { getYesNoDoNotKnowNotApplicableOptions } from '../../../utils/childProtectionNotificationUtils'
-import { getApplicationAnswers } from '../../../utils/getApplicationAnswers'
+import { memmMessages } from '../../../../lib/messages'
+import { getYesNoDoNotKnowNotApplicableOptions } from '../../../../utils/childProtectionNotificationUtils'
+import { getApplicationAnswers } from '../../../../utils/getApplicationAnswers'
 
 export const receptionSubSection = buildSubSection({
   id: 'memmReceptionSubSection',
