+++ libs/application/templates/national-agency-for-children-and-families/child-protection-notification/src/forms/adultProcuration/draftForm/reasonForNotificationSection/index.ts
@@ -1,5 +1,5 @@
 import { buildSection } from '@island.is/application/core'
-import { reasonForNotificationMessages } from '../../../lib/messages'
+import { reasonForNotificationMessages } from '../../../../lib/messages'
 import { reasonDescriptionSubSection } from './reasonDescriptionSubSection'
 import { reasonForNotificationSubSection } from './reasonForNotificationSubSection'
 import { reasonNotificationHistorySubSection } from './reasonNotificationHistorySubSection'
