+++ libs/application/templates/national-agency-for-children-and-families/child-protection-notification/src/forms/adultProcuration/draftForm/reasonForNotificationSection/reasonDescriptionSubSection.ts
@@ -5,7 +5,7 @@ import {
   buildTextField,
   YES,
 } from '@island.is/application/core'
-import { reasonForNotificationMessages } from '../../../lib/messages'
+import { reasonForNotificationMessages } from '../../../../lib/messages'
 
 export const reasonDescriptionSubSection = buildSubSection({
   id: 'reasonDescriptionSubSection',
