+++ libs/application/templates/national-agency-for-children-and-families/child-protection-notification/src/forms/adultProcuration/draftForm/reasonForNotificationSection/reasonForNotificationSubSection.ts
@@ -6,15 +6,15 @@ import {
   buildSelectField,
   buildSubSection,
 } from '@island.is/application/core'
-import { reasonForNotificationMessages } from '../../../lib/messages'
-import { getSelectedReasonForNotificationCategoryCodes } from '../../../utils/childProtectionNotificationUtils'
+import { reasonForNotificationMessages } from '../../../../lib/messages'
+import { getSelectedReasonForNotificationCategoryCodes } from '../../../../utils/childProtectionNotificationUtils'
 import {
   isUnborn,
   shouldShowBiggestConcernField,
   shouldShowReasonForNotificationSubCategoryDetails,
-} from '../../../utils/conditionUtils'
-import { RISK_TO_UNBORN } from '../../../utils/constants'
-import { getApplicationExternalData } from '../../../utils/getApplicationExternalData'
+} from '../../../../utils/conditionUtils'
+import { RISK_TO_UNBORN } from '../../../../utils/constants'
+import { getApplicationExternalData } from '../../../../utils/getApplicationExternalData'
 
 export const reasonForNotificationSubSection = buildSubSection({
   id: 'reasonForNotificationSubSection',
