+++ libs/application/templates/national-agency-for-children-and-families/child-protection-notification/src/forms/minor/draftForm/index.ts
@@ -1,12 +1,12 @@
 import { buildForm } from '@island.is/application/core'
 import { FormModes } from '@island.is/application/types'
 
-import { conclusionSection } from '../shared/conclusionSection'
+import { conclusionSection } from '../../shared/conclusionSection'
 import { overviewSection } from './overviewSection'
 import { minorFirstSection } from './firstSection'
 
-export const MinorForm = buildForm({
-  id: 'MinorForm',
+export const MinorDraftForm = buildForm({
+  id: 'MinorDraftForm',
   mode: FormModes.DRAFT,
   children: [minorFirstSection, overviewSection, conclusionSection],
 })
